Final cleanup and added phu: icon handling
